Refactor property weightings
